Version 1.4.5
Fix - Prices for the Bookable products from WooCommerce Bookings plugin were not converted to the default currency when 'Convert to Shop default currency (and set cart and checkout behavior separately)' option is selected for Shop Behaviour setting. This is fixed now.

Fix - The prices for the selected currency for the bookable products were shown as 0 on the Shop page when 'Add original price in shop' option is checked. This is fixed now.

Fix - A fatal error was shown on the Shop page for Variable products when visited as Wholesale customer added from Wholesale Customers For Woo plugin. This is fixed now.