Fix for #40
Added new else condition to check for products that do not have the currency set from our plugin & are using the default shop currency.

Tested for Shop behaviour set to "show prices in different currencies" and when cart/checkout behaviour is set to "Convert to currency of last product in cart" or "Convert to currency of first product in cart".

Fix #40